refactor(binding): move BindingMode to runtime-html
