+++ packages/__tests__/2-runtime/ast.integration.spec.ts
@@ -1,10 +1,10 @@
 import {
   AccessScopeExpression,
-  BindingMode,
   ConditionalExpression,
   LifecycleFlags
 } from '@aurelia/runtime';
 import {
+  BindingMode,
   IPlatform,
   LetBinding,
   PropertyBinding,
