+++ packages/__tests__/3-runtime-html/custom-elements.spec.ts
@@ -1,5 +1,4 @@
-import { BindingMode } from '@aurelia/runtime';
-import { CustomElement, ValueConverter } from '@aurelia/runtime-html';
+import { BindingMode, CustomElement, ValueConverter } from '@aurelia/runtime-html';
 import { assert, createFixture } from '@aurelia/testing';
 
 describe('3-runtime-html/custom-elements.spec.ts', function () {
