+++ packages/__tests__/3-runtime-html/decorator-watch.expression.spec.ts
@@ -1,5 +1,4 @@
-import { BindingMode } from '@aurelia/runtime';
-import { bindable, customAttribute, customElement, watch } from '@aurelia/runtime-html';
+import { BindingMode, bindable, customAttribute, customElement, watch } from '@aurelia/runtime-html';
 import { assert, createFixture, TestContext } from '@aurelia/testing';
 
 describe('3-runtime-html/decorator-watch.expression.spec.ts', function () {
