+++ packages/__tests__/3-runtime-html/integration.spec.ts
@@ -4,8 +4,6 @@ import {
 } from '@aurelia/kernel';
 import {
   BindingMode,
-} from '@aurelia/runtime';
-import {
   Aurelia,
   Controller,
   CustomElement,
