+++ packages/__tests__/3-runtime-html/spread.spec.ts
@@ -1,6 +1,5 @@
 import { Constructable } from '@aurelia/kernel';
-import { BindingMode } from '@aurelia/runtime';
-import { CustomAttribute, CustomElement, ICustomElementViewModel, INode } from '@aurelia/runtime-html';
+import { BindingMode, CustomAttribute, CustomElement, ICustomElementViewModel, INode } from '@aurelia/runtime-html';
 import { assert, createFixture } from '@aurelia/testing';
 
 // all the tests are using a common <my-input/> with a spreat on its internal <input/>
