+++ packages/__tests__/3-runtime-html/template-compiler.convention.spec.ts
@@ -1,7 +1,5 @@
 import {
   BindingMode,
-} from '@aurelia/runtime';
-import {
   ITemplateCompiler,
   IInstruction,
   InstructionType as TT,
