+++ packages/__tests__/3-runtime-html/template-compiler.spec.ts
@@ -11,13 +11,13 @@ import {
   parseExpression,
   AccessScopeExpression,
   BindingIdentifier,
-  BindingMode,
   DelegationStrategy,
   PrimitiveLiteralExpression,
   IExpressionParser,
 } from '@aurelia/runtime';
 import {
   bindable,
+  BindingMode,
   BindableDefinition,
   customAttribute,
   CustomAttribute,
