+++ packages/i18n/src/t/translation-renderer.ts
@@ -6,9 +6,9 @@ import {
   IExpressionParser,
   IObserverLocator,
   type IsBindingBehavior,
-  BindingMode,
 } from '@aurelia/runtime';
 import {
+  BindingMode,
   CommandType,
   IRenderer,
   renderer,
