+++ packages/router-lite/src/resources/load.ts
@@ -1,6 +1,5 @@
 import { IDisposable, IIndexable } from '@aurelia/kernel';
-import { BindingMode } from '@aurelia/runtime';
-import { customAttribute, bindable, ICustomAttributeViewModel, IEventDelegator, IEventTarget, INode, CustomElement } from '@aurelia/runtime-html';
+import { BindingMode, customAttribute, bindable, ICustomAttributeViewModel, IEventDelegator, IEventTarget, INode, CustomElement } from '@aurelia/runtime-html';
 
 import { IRouter } from '../router';
 import { IRouteContext } from '../route-context';
