+++ packages/router/src/resources/considered-active.ts
@@ -1,5 +1,4 @@
-import { BindingMode } from '@aurelia/runtime';
-import { customAttribute, bindable, ICustomAttributeViewModel } from '@aurelia/runtime-html';
+import { BindingMode, customAttribute, bindable, ICustomAttributeViewModel } from '@aurelia/runtime-html';
 
 @customAttribute('considered-active')
 export class ConsideredActiveCustomAttribute implements ICustomAttributeViewModel {
