+++ packages/router/src/resources/load.ts
@@ -1,6 +1,5 @@
 import { IDisposable, IEventAggregator } from '@aurelia/kernel';
-import { BindingMode } from '@aurelia/runtime';
-import { customAttribute, INode, bindable, CustomAttribute, ICustomAttributeViewModel } from '@aurelia/runtime-html';
+import { BindingMode, customAttribute, INode, bindable, CustomAttribute, ICustomAttributeViewModel } from '@aurelia/runtime-html';
 import { ILinkHandler } from './link-handler';
 import { IRouter, RouterNavigationEndEvent } from '../router';
 import { getConsideredActiveInstructions, getLoadIndicator } from './utils';
