+++ packages/runtime-html/src/binding-behaviors/binding-mode.ts
@@ -1,5 +1,6 @@
-import { BindingBehaviorInstance, BindingMode, LifecycleFlags } from '@aurelia/runtime';
+import { BindingBehaviorInstance, LifecycleFlags } from '@aurelia/runtime';
 import { bindingBehavior } from '../resources/binding-behavior';
+import { BindingMode } from '../binding/interfaces-bindings';
 
 import type { Scope } from '@aurelia/runtime';
 import type { PropertyBinding } from '../binding/property-binding';
