+++ packages/runtime-html/src/index.ts
@@ -115,6 +115,7 @@ export {
   type IsTwoWayPredicate,
 } from './attribute-mapper';
 export {
+  BindingMode,
   IAstBasedBinding,
   IBindingController,
 } from './binding/interfaces-bindings';
