+++ packages/runtime-html/src/resources/template-controllers/promise.ts
@@ -1,10 +1,11 @@
 import { Task, TaskAbortError, TaskStatus } from '@aurelia/platform';
 import { ILogger, nextId, onResolve, resolveAll } from '@aurelia/kernel';
-import { BindingMode, LifecycleFlags, Scope } from '@aurelia/runtime';
+import { LifecycleFlags, Scope } from '@aurelia/runtime';
 import { bindable } from '../../bindable';
 import { INode, IRenderLocation } from '../../dom';
 import { IPlatform } from '../../platform';
 import { IInstruction } from '../../renderer';
+import { BindingMode } from '../../binding/interfaces-bindings';
 import {
   Controller,
   ICustomAttributeController,
